Allow subscribe by tag without form
Updates the subscribe_email function to allow the ability to add a tag to subscriber without subscribing to form. ConvertKit API allows us to subscribe a user to a tag itself and, if the user is already a subscriber, adds the tag to the subscriber. This is useful when the user is already a subscriber and we simply want to add a tag to them rather than subscribing them to a form.